Screen Hopper
This commit turns HID Remapper into a USB input switcher. With it you
can connect a mouse/keyboard to two computers and switch between them
by dragging the mouse cursor from one screen to the other.

We do this by keeping an internal cursor position and sending absolute
cursor positions to the two computers (this way we know when to switch
from one to the other).

Because we need one more USB interface, HID Remapper's single/dual
Pico versions become dual/triple - the additional Pico is connected to
the other computer and receives the inputs it needs to send to it over
a serial link.